backEnd/services/auth-service/requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
sqlalchemy
psycopg2-binary
werkzeug
pyjwt
